DX-687 Map new payment details fields
New fields (default to empty string, if not available):

billerCode string BPAY Biller Code for the transaction

billerName string Name of the BPAY biller for the transaction

crn string BPAY CRN for the transaction

apcaNumber string 6 Digit APCA number for the initiating institution. The field is fixed-width and padded with leading zeros if applicable.

Co-Authored-By: Kam Rafiei <137843487+kambasiq@users.noreply.github.com>